backend/requirements.txt
# =================================================================
# NBA Analytics Backend - Python Dependencies
# =================================================================
# Updated: November 2025
# Python Version: 3.11+

# =================================================================
# Core FastAPI Framework
# =================================================================
fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.25.0,<0.32.0
pydantic>=2.5.3,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# =================================================================
# Database & ORM
# =================================================================
supabase>=2.7.4,<3.0.0
asyncpg>=0.29.0,<1.0.0
sqlalchemy>=2.0.25,<3.0.0

# =================================================================
# Async & Scheduling
# =================================================================
APScheduler>=3.10.4,<4.0.0
anyio>=4.2.0,<5.0.0

# =================================================================
# HTTP Requests & Web Scraping
# =================================================================
aiohttp>=3.9.1,<4.0.0
httpx>=0.27.0,<0.28.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.2,<5.0.0
lxml>=4.9.4,<5.0.0
selenium>=4.16.0,<5.0.0

# =================================================================
# Anti-Bot Protection & Stealth
# =================================================================
fake-useragent>=1.4.0,<2.0.0
cloudscraper>=1.2.71,<2.0.0
undetected-chromedriver>=3.5.4,<4.0.0
playwright>=1.40.0,<2.0.0
httpx-socks>=0.7.7,<1.0.0
python-socks>=2.4.3,<3.0.0
aiofiles>=23.2.1,<24.0.0

# =================================================================
# Data Processing & Analysis
# =================================================================
pandas>=2.1.4,<3.0.0
numpy>=1.26.2,<2.0.0
scipy>=1.11.4,<2.0.0

# =================================================================
# Date/Time & Timezone Support
# =================================================================
pytz>=2023.3.post1
python-dateutil>=2.8.2,<3.0.0

# =================================================================
# Environment & Configuration
# =================================================================
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.1,<7.0.0

# =================================================================
# NBA Data APIs
# =================================================================
nba-api>=1.4.1,<2.0.0

# =================================================================
# Logging & Monitoring  
# =================================================================
loguru>=0.7.2,<1.0.0
structlog>=23.2.0,<24.0.0

# =================================================================
# Security & Validation
# =================================================================
cryptography>=42.0.0,<47.0.0
certifi>=2023.11.17

# =================================================================
# Development & Testing (Optional)
# =================================================================
pytest>=7.4.4,<8.0.0
pytest-asyncio>=0.23.2,<1.0.0
black>=23.12.0,<25.0.0
isort>=5.13.2,<6.0.0
flake8>=7.0.0,<8.0.0

# =================================================================
# Production Extras
# =================================================================
gunicorn>=21.2.0,<22.0.0
redis>=5.0.1,<6.0.0
